*后端框架由原来的mybatisplus-spring-boot-starter 修改为MyBatis Plus Boot Starter *去除原来spring-mvc所使用的的mybatis-plus 统一修改为MyBatis Plus Boot Starter *修改因为修改框架导致的代码重构问题

+++ BlogWorkspace/summer-cat-common/src/main/java/com/summer/cat/util/Returns.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Returns {
-    public static Map<String, Object> mapOk(Object list, int total, String msg) {
+    public static Map<String, Object> mapOk(Object list, long total, String msg) {
         Map<String, Object> returnsMap = new HashMap<>(3);
         returnsMap.put(Constant.ReturnsMessage.TOTAL, total);
         returnsMap.put(Constant.ReturnsMessage.MESSAGE, msg);
